Revert "UBUNTU: (pre-stable): input: Support Clickpad devices in ClickZone mode"
BugLink: http://bugs.launchpad.net/bugs/669399

This reverts commit da04c34f6499177fb0cdafafa04af38c36b11639.

This patch was never adopted by upstream and is non-functional on
maverick in it's current incarnation (LP: #582809). The improper checks
for SYN_CAP_CLICKPAD(priv->ext_cap) instead of
SYN_CAP_CLICKPAD(priv->ext_cap_0c) in the maverick version of this patch
are also causing incorrect detection of a non-clickpad touchpad as a
clickpad one and applying the filters which remove the bottom area of
the touchpad. This revert only applies to maverick, lucid works properly
due to not having 5f57d67da87332a9a1ba8fa7a33bf0680e1c76e7 and this
commit was already dropped in natty.

Signed-off-by: Robert Hooker <robert.hooker@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>